move to psr-4 stucture and cleanup imports and cs
